add proper support for error bar plots via `ErrorBar` type
Several `newErrorBar` helper procs are defined, which allow easy
creation of an error bar, which can be assigned to the new `xs_err`
and `ys_err` fields of a Trace.

TODO: it would be nice to have to hand a color to the error
bar. Instead we might use the color of the `Trace.marker` when parsing
the `ErrorBar` to a `JsonNode`.