add `range` field to `Axis`
The `range` field allows to set the range of the `Axis`. This is
useful for example in case one sets a start and end bin range for a
histogram. Plotly drops empty bins from the calculation of the range
for its axes. To force it to show all bins, we can set the axis range
to the same value as the histogram bin range.