add different options to change histogram settings (bin size etc.) (#16)
* add different options to change histogram settings (bin size etc.)

Adds more options to customize histogram plots. It's now possible to
set different bin numbers, bin ranges, bin sizes, create cumulative
and normalized histograms.

Adds a `parseHistogramFields` func, which moves the parsing of the
histogram `Trace` variant out of the main `Trace` api func.

* add heatmap and candlestick to README

* fix autobinning if `nbins` given

If the number of maximum bins is given, the `autobin` field of plotly
needs to be set to true, instead of false, because the autobin
algorithm uses nbins as an input.

* add `range` field to `Axis`

The `range` field allows to set the range of the `Axis`. This is
useful for example in case one sets a start and end bin range for a
histogram. Plotly drops empty bins from the calculation of the range
for its axes. To force it to show all bins, we can set the axis range
to the same value as the histogram bin range.

* add example to showcase histogram settings, add to `all.nim`