guard saving of images behind check for --threads:on, fatal else
Puts the `show` and `saveImage` procs behind a check for
`--threads:on` compile flag.

If the user tries to save an image and compiles without thread
support, we quit the compilation procedure using fatal.

This introduces some violation of the DRY principle for the sake of
better error messages for the user at compile time.