make tests run in firefox on travis using virtual X server
This allows to also run the "save image" test, which runs plotly in
firefox and sends the data to our Nim program.

To achieve it we change from a docker based travis solution to using
choosenim and the firefox addon. In addition we replace Nim's
`openDefaultBrowser` by a version, which always appends a ` &` to the
opened URL to get non-blocking behaviour (which weirdly happends on
travis otherwise). We could alternatively modify the URL we open, but
this way the intent is clearer and more separated from the normal code.