Add ability to save plots as png, jpg, svg, webp (#18)
* add plotly/image_retrieve to receive image data from javascript

This submodule starts an async HTTP websocket server, which listens
for image data packages from the plotly javascript code running in the
browser. It decodes the packages and saves them as image files on disk.

* modify tmpl_html to allow injection of image save data

* modify `show` and `save` to inject js code to send img via websocket

* add convenience `saveImage` proc (replacement for show) and example

* add `websocket` requirement to nimble file

* add --threads:on to nimble test task

* fix `show` to only create websocket server, if filename given

* make sure we wait for the save image thread to join

* introduce channel to let server know when to close, poll manually

* add `withDebug` template to inject echos iff compiled with -d:DEBUG

Adds some visual noise to the code, if something breaks in the async /
threading part it's very useful to see where things break.

* move "Saving plot..." message to actual location of `writeFile`

* guard saving of images behind check for --threads:on, fatal else

Puts the `show` and `saveImage` procs behind a check for
`--threads:on` compile flag.

If the user tries to save an image and compiles without thread
support, we quit the compilation procedure using fatal.

This introduces some violation of the DRY principle for the sake of
better error messages for the user at compile time.

* remove fig12 from `all.nim`, add note to fig12 about compile checks

* change signature for `show` for "save image" case, filename 2nd arg

This allows us to make a distinction of the two procs `show`, which
are created if the user compiles without --threads:on for error
messages. Issues:
1. filename needs to be mandatory for the overload of the two `show`
to work, because otherwise they are ambiguous if filename is not given.
2. filename cannot be the last argument if it's mandatory (because
others before are not mandatory).

* add fig12 as individual test case compiled with --threads:on

* make tests run in firefox on travis using virtual X server

This allows to also run the "save image" test, which runs plotly in
firefox and sends the data to our Nim program.

To achieve it we change from a docker based travis solution to using
choosenim and the firefox addon. In addition we replace Nim's
`openDefaultBrowser` by a version, which always appends a ` &` to the
opened URL to get non-blocking behaviour (which weirdly happends on
travis otherwise). We could alternatively modify the URL we open, but
this way the intent is clearer and more separated from the normal code.