Support subplot creation at runtime via `createGrid` (#45)
* move impl of `plotly.nim` to `plotly_display` to make it importable

* add `createGrid` to create a grid subplot at runtime

`createGrid` returns a `Grid` object for which `[]`, `[]=` and `add`
are defined, so that the user may hand any `Plot[T]` object to the
`Grid`. Internally all plots are stored in a `seq[PlotJson]`.

* export `showImpl` plot to get `PlotJson` from grid

* fix JS backend

* only add plots, which have been assigned, i.e. not nil

The result of this is we can have e.g. a 2x2 grid of plots and only
assign 3 without a crash. However, if we only assign (0, 0), (0, 1),
and (1, 1), the plot at (1, 1) will be shifted to (1, 0)!

* add bounds check for `[]=` at grid index

* fix calculation of grid size if numPlotsPerRow > 0

Previously `numPlotsPerRow` caused the calculation of the # of rows
and columns to break, because we assumed `rows == 0` as a
placeholder.

Introduces `-1` as a special case for row or column, which means
"infer from the other dimension + nPlots".

* add creation and assignment of `Grid` based on row, col tuples

* remove left over JS related code from plotly_display

Since `parseTraces` is also useful on the JS backend, it was in
addition added to `plotly_js`, since either putting it into
`plotly_sugar` or creating a separate file for JS + C specific
functions seemed unreasonable.

* actually run the subplots example with nimble test

* fix assignment coord tuple in grid example using

* rename `showImpl` to `toPlotJson`

If the proc is exported its name should better reflect what it
actually does

* only import display in subplots if not JS target

`show` is only needed for the targets other than JS.
`hasThreadSupport` is defined here again to avoid having to import it
too from display.

* import types in plotly_js to get `parseTraces` working